Math now fully rendered

+++ rust/src/bin/scrape_chrome.rs
@@ -1,39 +1,36 @@
+use color_eyre::Result;
+use fizika::{
+    scrape_utils::{create_fizika_tab, fix_courses, get_links, process_tab},
+    MATH_NOT_RENDERED_COUNTER,
+};
+use headless_chrome::{Element, Tab};
+use inquire::Confirm;
+use itertools::Itertools;
+use select::document::Document;
 use std::{
     error::Error,
     fs::{create_dir, remove_dir_all, remove_file, File},
     io::Write,
     path::Path,
     sync::Arc,
-    thread::sleep,
-    time::Duration,
 };
-
-use color_eyre::Result;
-use fizika::scrape_utils::{create_fizika_tab, fix_courses, get_links, process_tab};
-use headless_chrome::Tab;
-use itertools::Itertools;
-use select::document::Document;
 use url::Url;
 
 fn main() -> Result<(), Box<dyn Error>> {
-    /* {
-        println!("This will delete every course. Type \"yes\" if you want to proceed");
+    let should_continue = Confirm::new("Do you want to proceed")
+        .with_default(true)
+        .with_help_message("This will delete the \"courses\" directory")
+        .prompt()?;
 
-        let stdin = stdin();
-        let mut iterator = stdin.lock().lines();
-        let line = iterator.next().unwrap().unwrap();
-
-        if line != "yes" {
-            println!("Quitting");
-            panic!();
-        }
-    } */
+    if !should_continue {
+        return Ok(());
+    }
 
     let (tab, _browser) = create_fizika_tab()?;
 
-    let document = get_html(Arc::clone(&tab));
+    let document = get_html_from_element(&get_element_from_tab(&tab));
 
-    let lines = get_links(&document)?;
+    let lines = get_links(&Document::from(document.as_str()))?;
     let mut line_file = File::create("./links.txt")?;
     let url = Url::parse(&tab.get_url())?;
     line_file.write_all(
@@ -58,17 +55,52 @@ fn main() -> Result<(), Box<dyn Error>> {
     for (pos, line) in lines.into_iter().skip(0).enumerate() {
         let new_address = url.join(&line)?;
         tab.navigate_to(&new_address.to_string())?;
-        tab.wait_until_navigated()?;
+        tab.wait_until_navigated().unwrap();
+
+        let document_element = get_element_from_tab(&tab);
+        let document_html = get_html_from_element(&document_element);
+
+        // wait for math render to complete
+        let promise = document_element.call_js_fn(
+            r#"async function wait_for_math() {
+                console.log("\"Wait for math\" function injected")
+                return new Promise((resolve, reject) => {
+                    setTimeout(() => {
+                        console.log("Math can not be rendered")
+                        reject(false)
+                    }, 10_000)
+
+                    MathJax.Hub.Register.StartupHook("End",function () {
+                        console.log("Math rendered")
+                        resolve(true);
+                    });
+                })
+        }"#,
+            vec![],
+            true,
+        )?;
+
+        match promise.value {
+            Some(value) => {
+                if !value.as_bool().unwrap() {
+                    unsafe { MATH_NOT_RENDERED_COUNTER += 1 }
+                }
+            }
+            None => unreachable!(),
+        }
 
         let dir_name = pages_dir.join(pos.to_string());
         create_dir(&dir_name)?;
         create_dir(dir_name.join("exercises"))?;
         let dir_name = dir_name.canonicalize()?;
 
-        let chapter_info = process_tab(get_html(Arc::clone(&tab)), dir_name.as_path(), pos)?;
+        let chapter_info = process_tab(
+            &Document::from(document_html.as_str()),
+            dir_name.as_path(),
+            pos,
+        )?;
 
         chapter_infos.push(chapter_info);
-        sleep(Duration::from_millis(500));
     }
 
     let chapter_info_dir = Path::new("chapter_infos.txt");
@@ -82,17 +114,22 @@ fn main() -> Result<(), Box<dyn Error>> {
 
     fix_courses()?;
 
+    unsafe {
+        println!("Math not rendered: {}", MATH_NOT_RENDERED_COUNTER);
+    }
+
     Ok(())
 }
 
-fn get_html(tab: Arc<Tab>) -> Document {
-    let html_value = tab
-        .find_element("html")
-        .unwrap()
-        .call_js_fn("function () { return this.outerHTML; }", false)
+fn get_element_from_tab<'a>(tab: &'a Arc<Tab>) -> Element<'a> {
+    tab.find_element("html").unwrap()
+}
+
+fn get_html_from_element(element: &Element) -> String {
+    let html_value = element
+        .call_js_fn("function () { return this.outerHTML; }", vec![], false)
         .unwrap();
 
     let course_html = html_value.value.unwrap();
-    let course_str = course_html.as_str().unwrap();
-    Document::from(course_str)
+    course_html.as_str().unwrap().to_string()
 }
